# api/openapi.yaml
openapi: 3.0.3
info:
  title: Warrant API
  description: REST API for Warrant ticketing (orgs, projects, tickets, reviews, queue).
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Local
paths:
...
  /projects/{projectID}:
    get:
      operationId: GetProject
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
    patch:
      operationId: UpdateProject
      summary: Update project (e.g. set status to active or closed)
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        "400":
          description: Bad request (e.g. invalid status)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
  /projects/{projectID}/tickets:
    get:
      operationId: ListTickets
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
    post:
      operationId: CreateTicket
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTicketRequest'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
  /tickets/{ticketID}:
    get:
      operationId: GetTicket
      parameters:
        - name: ticketID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
    patch:
      operationId: UpdateTicket
      parameters:
        - name: ticketID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTicketRequest'
      responses:
        "204":
          description: No content
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
...
  /tickets/{ticketID}/transitions:
    post:
      operationId: TransitionTicket
      summary: Transition ticket state (e.g. start, submit, approve, reject)
      security:
        - BearerAuth: []
      parameters:
        - name: ticketID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransitionRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad request (e.g. trigger required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    StructuredError:
      type: object
      required: [error, code, retriable]
      properties:
        error:
          type: string
        code:
          type: string
          enum: [lease_expired, unauthorized, not_found, conflict, invalid_input, forbidden, internal, not_implemented, project_closed]
        retriable:
          type: boolean
...
    TransitionRequest:
      type: object
      required: [trigger]
      properties:
        trigger:
          type: string
        actor_id:
          type: string
        actor:
          type: string
          enum: [human, agent, system]
        payload:
          type: object
...
    Project:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        name:
          type: string
        slug:
          type: string
        repo_url:
          type: string
        tech_stack:
          type: array
          items:
            type: string
        context_pack:
          type: object
        status:
          type: string
          enum: [active, closed]
          description: active (default) or closed; list endpoints default to active only.
        created_at:
          type: string
          format: date-time
    UpdateProjectRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, closed]
...
    Objective:
      type: object
      properties:
        description:
          type: string
        success_criteria:
          type: array
          items:
            type: string
        acceptance_test:
          type: string
    TicketContext:
      type: object
      properties:
        relevant_files:
          type: array
          items:
            type: string
        constraints:
          type: array
          items:
            type: string
        prior_attempts:
          type: array
          items:
            type: object
        human_answers:
          type: array
          items:
            type: string
    Ticket:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [task, bug, spike, review]
        priority:
          type: integer
          minimum: 0
          maximum: 3
        state:
          type: string
          enum: [pending, claimed, executing, awaiting_review, done, blocked, needs_human, failed]
        version:
          type: integer
        objective:
          $ref: '#/components/schemas/Objective'
        ticket_context:
          $ref: '#/components/schemas/TicketContext'
        inputs:
          type: object
        outputs:
          type: object
        depends_on:
          type: array
          items:
            type: string
        assigned_to:
          type: string
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateTicketRequest:
      type: object
      required: [title, created_by]
      properties:
        title:
          type: string
        type:
          type: string
          enum: [task, bug, spike, review]
        priority:
          type: integer
          minimum: 0
          maximum: 3
        created_by:
          type: string
        depends_on:
          type: array
          items:
            type: string
        objective:
          $ref: '#/components/schemas/Objective'
        ticket_context:
          $ref: '#/components/schemas/TicketContext'
        idempotency_key:
          type: string
          description: Optional. Duplicate requests with the same key return the existing ticket.
    UpdateTicketRequest:
      type: object
      properties:
        depends_on:
          type: array
          items:
            type: string
...